# Added value type serialization support