--- a/src/System.Text.Json/src/System/Text/Json/Serialization/ReflectionEmitMaterializer.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/ReflectionEmitMaterializer.cs
@@ -15,22 +15,34 @@ namespace System.Text.Json.Serialization
         public override JsonClassInfo.ConstructorDelegate CreateConstructor(Type type)
         {
             Debug.Assert(type != null);
-
             ConstructorInfo realMethod = type.GetConstructor(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance, binder: null, Type.EmptyTypes, modifiers: null);
-            if (realMethod == null)
-            {
+
+            if (realMethod == null && !type.IsValueType)
                 return null;
-            }
 
             var dynamicMethod = new DynamicMethod(
-                realMethod.Name,
-                type,
+                ConstructorInfo.ConstructorName,
+                typeof(object),
                 Type.EmptyTypes,
                 typeof(ReflectionEmitMaterializer).Module,
                 skipVisibility: true);
 
             ILGenerator generator = dynamicMethod.GetILGenerator();
-            generator.Emit(OpCodes.Newobj, realMethod);
+
+            if (realMethod == null)
+            {
+                LocalBuilder local = generator.DeclareLocal(type);
+
+                generator.Emit(OpCodes.Ldloca_S, local);
+                generator.Emit(OpCodes.Initobj, type);
+                generator.Emit(OpCodes.Ldloc, local);
+                generator.Emit(OpCodes.Box, type);
+            }
+            else
+            {
+                generator.Emit(OpCodes.Newobj, realMethod);
+            }
+
             generator.Emit(OpCodes.Ret);
 
             return (JsonClassInfo.ConstructorDelegate)dynamicMethod.CreateDelegate(typeof(JsonClassInfo.ConstructorDelegate));