--- a/src/System.Text.Json/src/System/Text/Json/Serialization/ReflectionEmitMaterializer.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/ReflectionEmitMaterializer.cs
@@ -18,7 +18,9 @@ namespace System.Text.Json.Serialization
             ConstructorInfo realMethod = type.GetConstructor(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance, binder: null, Type.EmptyTypes, modifiers: null);
 
             if (realMethod == null && !type.IsValueType)
+            {
                 return null;
+            }
 
             var dynamicMethod = new DynamicMethod(
                 ConstructorInfo.ConstructorName,