This pull request add value type serialization support, but it works only for properties. Therefore, it fixes #36240 partly. Opened #36505 to track field serialization.

Open questions:
* Should a constructor be generated for primitive types or not? `ReflectionMaterializer` does it for all types, but `ReflectionEmitMaterializer` didn't previously.

/cc @ahsonkhan @steveharter