@steveharter 2019-04-01
> Should a constructor be generated for primitive types or not? ReflectionMaterializer does it for all types, but ReflectionEmitMaterializer didn't previously.

The reason for ReflectionEmitMaterializer  is performance (and without boxing), so base decisions on that...

Also currently "data types" (Decimal, int, etc) are created explicitly in JsonValueConverter.TryRead. These are "single value" data types. So this commit will be useful for value types that are not used as "single-value data types".