@YohDeadfall 2019-04-12
Removed the dirty trick with `Unsafe.As<T>` and added member access generator which doesn't emit IL. This is what @GrabYourPitchforks asked in #24390.