@GrabYourPitchforks 2019-04-13
This isn't quite what I had in mind with that proposal. Additionally, is there any way to do this without use of Unsafe APIs? Putting those into a serializer / deserializer makes me extraordinarily uncomfortable.