@YohDeadfall 2019-04-13
Sure, it's possible, but there is a cost in complexity. A stack frame should store a wrapper around the struct in the `ReturnValue` property or a class instance for reference types. For example see `ByReference<T>` and how it's used in npgsql/npgsql#2402.

Speaking about your proposal, how do you plan to implement property accessors ?