@GrabYourPitchforks 2019-04-13
I had planned on using a combination of a force-JIT flag and _calli_, which should be lighter-weight than bouncing through multiple delegate invocations. But importantly, the complexity of the unsafe code would be wholly encapsulated within coreclr (where we could keep it in sync with how the runtime works) rather than spread across non-coreclr packages which would have to be serviced independently if something in the runtime changes or if a bug were discovered.

(This shouldn't affect your PR much. It just means that if that other issue does make its way in at some future point then this code can be refactored to ride atop it, which should get a measurable perf boost.)

If the alternative for not using `Unsafe.Unbox` is added complexity, I'll not push back. :) Thanks for your responses!