@YohDeadfall 2019-04-14
@GrabYourPitchforks Totally agree that such a kind of stuff should be moved off. I don't like the chained delegates I introduced, but there is no other way for now except code generation. So I'm looking forward to `ReflectionServices` (:

@steveharter Rebased and squashed commits.