@steveharter 2019-04-23
cc @rynowak @JamesNK any concerns with supporting struct at the highest level? (the struct must still have public properties - fields not supported with this PR).

For areas that expose a top-level type, like in extensibilty scenarios, this means we will want to pass the type argument by reference (`ref` or `in` modifier) to allow the struct to be modified and\or to avoid unnecessary copies. Passing a reference type with `ref` or `in` is allowed, so we should be OK syntax-wise.