@ahsonkhan 2019-04-24
Imo, it could be difficult for customers to reason about the functionality with conditional support like this, so we should avoid piecemeal support of the struct feature. Today, the serializer supports classes. If we want to support structs, we should make sure that nested/custom structs (structs within reference types, reference types within structs, etc.) are supported, potentially with a story for fields/default values/ctors.