@rynowak 2019-04-24
Are there major blockers to lighting up value types all over the stack (even if it's slightly ineffient? requires boxing, etc). 

I'd like to say yes, to having struct support, but it's hard to explain to people why it would only work as a top-level object.