@YohDeadfall 2019-04-24
I've opened dotnet/coreclr#24189 to get some clarifications about an AOT friendly way to get the underlying value of an nullable, but for now it's possible to do the same trick which exists in Npgsql (see npgsql/npgsql#1773 for example). I think that it's okay because the current implementation doesn't support AOT out from the box.

So I can invert the code and make two final property infos, one for reference types and another one for value types. Nullable handling logic will be implemented via a delegate and a method to take an underlying value only if `TDeclaredProperty` is `Nullable<T>`. This path will be slower than non-nullable due to a delegate invocation, but the code will be clean and there will be no perf regression in other places. For example, see npgsql/npgsql#2402.