@steveharter 2019-04-25
cc @JeremyKuhne has a scenario where we have a top-level struct with public properties (actually just our JsonElement).