@steveharter 2019-04-26
@ahsonkhan @rynowak any final thoughts? I recommend supporting this. The other 3 serializers I tested (Json.Net, Jil, Utf8Json) support structs at top-level.

The perf hit on end-to-end is within the margin of error (simple POCO type with 10 properties. Struct performance is the same.

Before:
```
|           Method |      Mean |      Error |     StdDev |    Median |       Min |       Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|----------------- |----------:|-----------:|-----------:|----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
| System.Text.Json |  1.937 us |  0.0457 us |  0.0526 us |  1.930 us |  1.857 us |  2.067 us |      0.1210 |           - |           - |               512 B |
```

After:
```
class
|           Method |      Mean |     Error |    StdDev |    Median |       Min |       Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|----------------- |----------:|----------:|----------:|----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
| System.Text.Json |  1.965 us | 0.0393 us | 0.0386 us |  1.966 us |  1.913 us |  2.021 us |      0.1211 |           - |           - |               512 B |