@YohDeadfall 2019-05-04
@ahsonkhan Is there a documentation about bots? Can I ask it to run a pipeline?