@ahsonkhan 2019-05-04
> @ahsonkhan Is there a documentation about bots? Can I ask it to run a pipeline?

Yep, see this: 
https://github.com/dotnet/corefx/blob/648d0724c2b9224cace4b1e179fc20904fe27aa4/Documentation/project-docs/pullrequest-builds.md#available-builds