@ahsonkhan 2019-05-06
> The perf hit on end-to-end is within the margin of error (simple POCO type with 10 properties. Struct performance is the same.

Can you please verify that these benchmarks perf doesn't regress noticeably here as well?
https://github.com/dotnet/performance/pull/456