@YohDeadfall 2019-05-06
> Can you please verify that these benchmarks perf doesn't regress noticeably here as well?

I will try to do this too tomorrow. Unfortunately, I'm getting `OutOfMemoryException`s while building Core Fx as described in [the docs](https://github.com/dotnet/performance/blob/master/docs/benchmarking-workflow-corefx.md#corefx-prerequisites).

Can I mention @dotnet-bot here and ask him to run some checks? As I know it's possible in Core CLR and Core Fx Lab.