@ahsonkhan 2019-05-07
> Can I mention @dotnet-bot here and ask him to run some checks? As I know it's possible in Core CLR and Core Fx Lab.

I don't think so. What types of checks were you looking for? We don't have any perf validation CI legs.

> I will try to do this too tomorrow. Unfortunately, I'm getting `OutOfMemoryException`s while building Core Fx as described in [the docs](https://github.com/dotnet/performance/blob/master/docs/benchmarking-workflow-corefx.md#corefx-prerequisites).

@steveharter, would you be able to help run the changes in this PR against the perf tests from dotnet/performance#456 and compare before/after like you did previously?