@ahsonkhan 2019-05-19
> I am wondering what's causing the 20-40% regression on certain cases. I would have thought ToString and ToBytes to behave the same way in terms of perf, so the fact that they don't might point to perf results not being consistent.

I re-tried running the benchmarks and it looks like the performance is reasonable. I commented out some of the config that reduces accuracy in the interest of test execution time:
https://github.com/dotnet/performance/blob/778aa84d99344fcdf73b8c8cff3ff6dee9e5abf9/src/harness/BenchmarkDotNet.Extensions/RecommendedConfig.cs#L17-L20

**IndexViewModel**

``` ini

BenchmarkDotNet=v0.11.3.1003-nightly, OS=Windows 10.0.18362
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-preview4-010907
  [Host]     : .NET Core 3.0.0-preview6-27629-07 (CoreCLR 4.6.27624.71, CoreFX 4.700.19.22902), 64bit RyuJIT
  Job-BMHKSV : .NET Core ? (CoreCLR 3.0.19.26472, CoreFX 4.700.19.26901), 64bit RyuJIT

Runtime=Core  Toolchain=CoreRun  

```

**Before:**

|                   Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    DeserializeFromString | 53.20 us | 0.5569 us | 0.4650 us | 53.26 us | 52.67 us | 54.38 us |      8.3008 |      1.0376 |           - |            34.14 KB |
| DeserializeFromUtf8Bytes | 48.68 us | 0.8121 us | 0.7596 us | 48.54 us | 47.79 us | 50.33 us |      5.3101 |           - |           - |            21.85 KB |
|    DeserializeFromStream | 49.86 us | 0.9864 us | 1.0129 us | 49.77 us | 48.87 us | 52.41 us |      5.3101 |           - |           - |            21.92 KB |

**After:**

|                   Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    DeserializeFromString | 54.97 us | 1.6263 us | 1.4417 us | 54.46 us | 53.47 us | 58.98 us |      8.3008 |      1.0376 |           - |            34.14 KB |
| DeserializeFromUtf8Bytes | 47.92 us | 0.6944 us | 0.6155 us | 47.74 us | 46.84 us | 49.11 us |      5.3101 |           - |           - |            21.85 KB |
|    DeserializeFromStream | 50.36 us | 0.4107 us | 0.3430 us | 50.36 us | 49.76 us | 50.82 us |      5.3101 |           - |           - |            21.92 KB |

**Before:**

|               Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    SerializeToString | 39.26 us | 0.6044 us | 0.5653 us | 39.19 us | 38.65 us | 40.48 us |      6.0425 |           - |           - |             25568 B |
| SerializeToUtf8Bytes | 36.71 us | 0.8467 us | 0.7920 us | 36.38 us | 35.85 us | 38.76 us |      3.0518 |           - |           - |             13016 B |
|    SerializeToStream | 34.84 us | 0.5173 us | 0.4319 us | 34.80 us | 34.27 us | 35.82 us |      0.0610 |           - |           - |               432 B |

**After:**

|               Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    SerializeToString | 39.60 us | 0.6869 us | 0.6090 us | 39.47 us | 38.69 us | 41.07 us |      6.0425 |           - |           - |             25568 B |
| SerializeToUtf8Bytes | 37.43 us | 0.4864 us | 0.4550 us | 37.37 us | 36.78 us | 38.33 us |      3.0518 |           - |           - |             13016 B |
|    SerializeToStream | 35.93 us | 0.6962 us | 0.9294 us | 35.61 us | 34.98 us | 37.97 us |      0.0610 |           - |           - |               432 B |